Refactor map stores
